requirements.txt
google-generativeai>=0.8.0
psutil>=5.9.8
python-dotenv>=1.0.1
